1. Create aggregateHolidayBudget method in Holiday class 2. Add controller and service layer logic. 3. Create addHolidayMaker and removeHolidayMaker method in Holiday class

+++ src/main/java/com/example/holidayplanner/budget/Budget.java
@@ -7,6 +7,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.FieldType;
 import org.springframework.data.mongodb.core.mapping.MongoId;
 
+
 @Data
 @Document(collection = "Budgets")
 public class Budget {
@@ -19,17 +20,16 @@ public class Budget {
     private User user;
 
     @JsonProperty
-    private int budgetUpperLimit;
+    private double budgetUpperLimit;
 
     @JsonProperty
-    private int budgetLowerLimit;
+    private double budgetLowerLimit;
 
 
 
-    public Budget(User user, int budgetUpperLimit, int budgetLowerLimit) {
+    public Budget(User user, double budgetUpperLimit, double budgetLowerLimit) {
         this.user = user;
         this.budgetUpperLimit = budgetUpperLimit;
         this.budgetLowerLimit = budgetLowerLimit;
     }
-
 }
